chore: 타입 리팩토링 및 위치 변경
