+++ src/contexts/PaymentStatusSearchInfoContext.tsx
@@ -1,5 +1,6 @@
 import { createContext, Dispatch, ReactNode, SetStateAction, useState } from "react";
-import { PaginationModelType, SearchVariantType } from "@/types/entities/search";
+import { PaginationModelType } from "@/types/entities/common";
+import { SearchVariantType } from "@/types/entities/member";
 
 type PaymentStatusSearchInfoContextProviderPropsType = {
   children: ReactNode;
