+++ src/types/dtos/member.ts
@@ -1,32 +1,5 @@
 import { MemberInfoType, StatusType } from "@/types/entities/member";
-
-export type PaginationElementType = {
-  pageable: {
-    pageNumber: number;
-    pageSize: number;
-    sort: {
-      empty: boolean;
-      sorted: boolean;
-      unsorted: boolean;
-    };
-    offset: number;
-    paged: boolean;
-    unpaged: boolean;
-  };
-  totalPages: number;
-  totalElements: number;
-  last: boolean;
-  size: number;
-  number: number;
-  sort: {
-    empty: boolean;
-    sorted: boolean;
-    unsorted: boolean;
-  };
-  numberOfElements: number;
-  first: boolean;
-  empty: boolean;
-};
+import { PaginationElementType } from "./common";
 
 export type AllMemberListResponseDtoType = {
   content: MemberInfoType[];
@@ -46,23 +19,3 @@ export type PaymentStatusMemberListResponseDtoType = {
 } & PaginationElementType;
 
 export type MemberPaymentStatusResponseDtoType = StatusType;
-
-export type DetailMemberInfoType = {
-  name: string;
-  studentId: string;
-  phone: string;
-  departmentName: string;
-  email: string;
-  discordUsername: string;
-  nickname: string;
-};
-
-export type DetailMemberModalInfoType = {
-  "이름": DetailMemberInfoType["name"];
-  "학번": DetailMemberInfoType["studentId"];
-  "전화번호": DetailMemberInfoType["phone"];
-  "소속 학과": DetailMemberInfoType["departmentName"];
-  "이메일": DetailMemberInfoType["email"];
-  "디스코드 사용자명": DetailMemberInfoType["discordUsername"];
-  "디스코드 별명": DetailMemberInfoType["nickname"];
-};
