+++ src/types/entities/common.ts
