+++ src/types/entities/member.ts
@@ -8,6 +8,8 @@ export type ManagementVariant =
   | "couponProvision"
   | "couponProvisionMembers";
 
+export type EditMemberInfoType = Exclude<MemberInfoType, "requirement">;
+
 export type MemberInfoType = {
   memberId: number;
   studentId: string;
@@ -53,4 +55,24 @@ export type SearchVariantType = MembersSearchVariantType[number] | null;
 export type SearchInfoType = {
   text: string;
   variant: SearchVariantType;
+};
+
+export type DetailMemberInfoType = {
+  name: string;
+  studentId: string;
+  phone: string;
+  departmentName: string;
+  email: string;
+  discordUsername: string;
+  nickname: string;
+};
+
+export type DetailMemberModalInfoType = {
+  "이름": DetailMemberInfoType["name"];
+  "학번": DetailMemberInfoType["studentId"];
+  "전화번호": DetailMemberInfoType["phone"];
+  "소속 학과": DetailMemberInfoType["departmentName"];
+  "이메일": DetailMemberInfoType["email"];
+  "디스코드 사용자명": DetailMemberInfoType["discordUsername"];
+  "디스코드 별명": DetailMemberInfoType["nickname"];
 };
